cleaned up a bunch of circular dependencies
