+++ lib/wrappers/commonEventInterfaces.ts
@@ -1,53 +1,6 @@
 import { BigNumber } from "bignumber.js";
 import { Address, Hash } from "../commonTypes";
 
-export interface NewProposalEventResult {
-  _numOfChoices: BigNumber;
-  _paramsHash: Hash;
-  /**
-   * indexed
-   */
-  _proposalId: Hash;
-  _proposer: Address;
-  _avatar: Address;
-}
-
-// TODO: include _avatar?
-/**
- * fired by voting machines
- */
-export interface VotingMachineExecuteProposalEventResult {
-  /**
-   * the vote choice that won.
-   */
-  _decision: BigNumber;
-  /**
-   * indexed
-   */
-  _proposalId: Hash;
-  /**
-   * The total reputation in the DAO at the time the proposal was executed
-   */
-  _totalReputation: BigNumber;
-}
-
-// TODO: include _avatar?
-export interface VoteProposalEventResult {
-  /**
-   * indexed
-   */
-  _proposalId: Hash;
-  _reputation: BigNumber;
-  /**
-   * The choice of vote
-   */
-  _vote: number;
-  /**
-   * indexed
-   */
-  _voter: Address;
-}
-
 export interface ProposalDeletedEventResult {
   /**
    * indexed
